Add Simulation component and refactor SimulationPage
- Introduced a new Simulation component to handle the simulation logic and UI.
- Removed redundant code from SimulationPage by integrating the Simulation component.
- Added images for AI and User avatars.
- Ensured proper state management and question handling within the new Simulation component.
