+++ src/Components/CaseStudy/Simulator/ChatInterface.jsx
